// Frontend/redaplus/src/types/videoContent.ts
export interface IVideoFFProbeData {
  format: {
    size: string;
    tags: {
      major_brand: string;
      creation_time: string;
      minor_version: string;
      compatible_brands: string;
    };
    bit_rate: string;
    duration: string;
    filename: string;
    nb_streams: number;
    start_time: string;
    format_name: string;
    nb_programs: number;
    probe_score: number;
    format_long_name: string;
  };
  streams: [
    {
      refs: 1;
      tags: {
        language: 'und';
        vendor_id: '[0][0][0][0]';
        handler_name: 'ISO Media file produced by Google Inc. Created on: 11/28/2021.';
        creation_time: '2021-11-28T15:11:37.000000Z';
      };
      index: 0;
      level: 31;
      width: number;
      height: number;
      is_avc: 'true';
      pix_fmt: 'yuv420p';
      profile: 'Main';
      bit_rate: '689879';
      duration: '524.466667';
      codec_tag: '0x31637661';
      nb_frames: '15734';
      start_pts: 0;
      time_base: '1/15360';
      codec_name: 'h264';
      codec_type: 'video';
      start_time: '0.000000';
      coded_width: 1280;
      color_range: 'tv';
      color_space: 'bt709';
      disposition: {
        dub: 0;
        forced: 0;
        lyrics: 0;
        comment: 0;
        default: 1;
        karaoke: 0;
        original: 0;
        attached_pic: 0;
        clean_effects: 0;
        visual_impaired: 0;
        hearing_impaired: 0;
        timed_thumbnails: 0;
      };
      duration_ts: 8055808;
      coded_height: 720;
      has_b_frames: 1;
      r_frame_rate: '30/1';
      avg_frame_rate: '30/1';
      color_transfer: 'bt709';
      chroma_location: 'left';
      closed_captions: 0;
      codec_long_name: 'H.264 / AVC / MPEG-4 AVC / MPEG-4 part 10';
      color_primaries: 'bt709';
      nal_length_size: '4';
      codec_tag_string: 'avc1';
      bits_per_raw_sample: '8';
      sample_aspect_ratio: '1:1';
      display_aspect_ratio: '16:9';
    },
    {
      tags: {
        language: 'eng';
        vendor_id: '[0][0][0][0]';
        handler_name: 'ISO Media file produced by Google Inc. Created on: 11/28/2021.';
        creation_time: '2021-11-28T15:11:37.000000Z';
      };
      index: 1;
      profile: 'LC';
      bit_rate: '128005';
      channels: 2;
      duration: '524.469116';
      codec_tag: '0x6134706d';
      nb_frames: '22587';
      start_pts: 0;
      time_base: '1/44100';
      codec_name: 'aac';
      codec_type: 'audio';
      sample_fmt: 'fltp';
      start_time: '0.000000';
      disposition: {
        dub: 0;
        forced: 0;
        lyrics: 0;
        comment: 0;
        default: 1;
        karaoke: 0;
        original: 0;
        attached_pic: 0;
        clean_effects: 0;
        visual_impaired: 0;
        hearing_impaired: 0;
        timed_thumbnails: 0;
      };
      duration_ts: 23129088;
      sample_rate: '44100';
      r_frame_rate: '0/0';
      avg_frame_rate: '0/0';
      channel_layout: 'stereo';
      bits_per_sample: 0;
      codec_long_name: 'AAC (Advanced Audio Coding)';
      codec_tag_string: 'mp4a';
    },
  ];
}